Reorganize PaymentRow* files
