+++ frontend/utils/time.js
@@ -1,7 +1,5 @@
 'use strict'
 
-const locale = navigator.languages ? navigator.languages[0] : navigator.language
-
 export const MINS_MILLIS = 60000
 export const HOURS_MILLIS = 60 * MINS_MILLIS
 export const DAYS_MILLIS = 24 * HOURS_MILLIS
@@ -50,5 +48,6 @@ export function humanDate (
   datems = Date.now(),
   opts = { month: 'short', day: 'numeric' }
 ) {
+  const locale = navigator.languages ? navigator.languages[0] : navigator.language
   return new Date(datems).toLocaleDateString(locale, opts)
 }
